+++ sdk/search/search-documents/src/search/api/operations.ts
@@ -101,6 +101,7 @@ export async function _autocompletePostDeserialize(
 
   return autocompleteResultDeserializer(result.body);
 }
+
 /** Autocompletes incomplete query terms based on input text and matching terms in the index. */
 export async function autocompletePost(
   context: Client,
@@ -171,6 +172,7 @@ export async function _autocompleteGetDeserialize(
 
   return autocompleteResultDeserializer(result.body);
 }
+
 /** Autocompletes incomplete query terms based on input text and matching terms in the index. */
 export async function autocompleteGet(
   context: Client,
@@ -228,6 +230,7 @@ export async function _indexDeserialize(
 
   return indexDocumentsResultDeserializer(result.body);
 }
+
 /** Sends a batch of document write actions to the index. */
 export async function index(
   context: Client,
@@ -297,6 +300,7 @@ export async function _suggestPostDeserialize(
 
   return suggestDocumentsResultDeserializer(result.body);
 }
+
 /** Suggests documents in the index that match the given partial query text. */
 export async function suggestPost(
   context: Client,
@@ -364,6 +368,7 @@ export async function _suggestGetDeserialize(
 
   return suggestDocumentsResultDeserializer(result.body);
 }
+
 /** Suggests documents in the index that match the given partial query text. */
 export async function suggestGet(
   context: Client,
@@ -427,6 +432,7 @@ export async function _getDocumentDeserialize(
 
   return lookupDocumentDeserializer(result.body);
 }
+
 /** Retrieves a document from the index. */
 export async function getDocument(
   context: Client,
@@ -546,6 +552,7 @@ export async function _searchPostDeserialize(
 
   return searchDocumentsResultDeserializer(result.body);
 }
+
 /** Searches for documents in the index. */
 export async function searchPost(
   context: Client,
@@ -650,6 +657,7 @@ export async function _searchGetDeserialize(
 
   return searchDocumentsResultDeserializer(result.body);
 }
+
 /** Searches for documents in the index. */
 export async function searchGet(
   context: Client,
@@ -700,6 +708,7 @@ export async function _getDocumentCountDeserialize(result: PathUncheckedResponse
 
   return result.body;
 }
+
 /** Queries the number of documents in the index. */
 export async function getDocumentCount(
   context: Client,