+++ sdk/search/search-documents/src/searchIndex/api/searchIndexContext.ts
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 import { logger } from "../../logger.js";
+import pkgJson from "@azure/search-documents/package.json" with { type: "json" };
 import { KnownVersions } from "../../models/models.js";
 import type { Client, ClientOptions } from "@azure-rest/core-client";
 import { getClient } from "@azure-rest/core-client";
@@ -27,7 +28,7 @@ export function createSearchIndex(
 ): SearchIndexContext {
   const endpointUrl = options.endpoint ?? String(endpointParam);
   const prefixFromOptions = options?.userAgentOptions?.userAgentPrefix;
-  const userAgentInfo = `azsdk-js-search-documents/13.1.0-beta.2`;
+  const userAgentInfo = `azsdk-js-search-documents/${pkgJson.version}`;
   const userAgentPrefix = prefixFromOptions
     ? `${prefixFromOptions} ${userAgentInfo}`
     : `${userAgentInfo}`;