+++ sdk/search/search-documents/src/searchIndexer/api/operations.ts
@@ -107,6 +107,7 @@ export async function _resetSkillsDeserialize(result: PathUncheckedResponse): Pr
 
   return;
 }
+
 /** Reset an existing skillset in a search service. */
 export async function resetSkills(
   context: Client,
@@ -165,6 +166,7 @@ export async function _createSkillsetDeserialize(
 
   return searchIndexerSkillsetDeserializer(result.body);
 }
+
 /** Creates a new skillset in a search service. */
 export async function createSkillset(
   context: Client,
@@ -223,6 +225,7 @@ export async function _getSkillsetsDeserialize(
 
   return _listSkillsetsResultDeserializer(result.body);
 }
+
 /** List all skillsets in a search service. */
 export function getSkillsets(
   context: Client,
@@ -287,6 +290,7 @@ export async function _getSkillsetDeserialize(
 
   return searchIndexerSkillsetDeserializer(result.body);
 }
+
 /** Retrieves a skillset in a search service. */
 export async function getSkillset(
   context: Client,
@@ -343,6 +347,7 @@ export async function _deleteSkillsetDeserialize(result: PathUncheckedResponse):
 
   return;
 }
+
 /** Deletes a skillset in a search service. */
 export async function deleteSkillset(
   context: Client,
@@ -407,6 +412,7 @@ export async function _createOrUpdateSkillsetDeserialize(
 
   return searchIndexerSkillsetDeserializer(result.body);
 }
+
 /** Creates a new skillset in a search service or updates the skillset if it already exists. */
 export async function createOrUpdateSkillset(
   context: Client,
@@ -464,6 +470,7 @@ export async function _getIndexerStatusDeserialize(
 
   return searchIndexerStatusDeserializer(result.body);
 }
+
 /** Returns the current status and execution history of an indexer. */
 export async function getIndexerStatus(
   context: Client,
@@ -521,6 +528,7 @@ export async function _createIndexerDeserialize(
 
   return searchIndexerDeserializer(result.body);
 }
+
 /** Creates a new indexer. */
 export async function createIndexer(
   context: Client,
@@ -579,6 +587,7 @@ export async function _getIndexersDeserialize(
 
   return _listIndexersResultDeserializer(result.body);
 }
+
 /** Lists all indexers available for a search service. */
 export function getIndexers(
   context: Client,
@@ -643,6 +652,7 @@ export async function _getIndexerDeserialize(
 
   return searchIndexerDeserializer(result.body);
 }
+
 /** Retrieves an indexer definition. */
 export async function getIndexer(
   context: Client,
@@ -699,6 +709,7 @@ export async function _deleteIndexerDeserialize(result: PathUncheckedResponse): 
 
   return;
 }
+
 /** Deletes an indexer. */
 export async function deleteIndexer(
   context: Client,
@@ -763,6 +774,7 @@ export async function _createOrUpdateIndexerDeserialize(
 
   return searchIndexerDeserializer(result.body);
 }
+
 /** Creates a new indexer or updates an indexer if it already exists. */
 export async function createOrUpdateIndexer(
   context: Client,
@@ -818,6 +830,7 @@ export async function _runIndexerDeserialize(result: PathUncheckedResponse): Pro
 
   return;
 }
+
 /** Runs an indexer on-demand. */
 export async function runIndexer(
   context: Client,
@@ -877,6 +890,7 @@ export async function _resetDocumentsDeserialize(result: PathUncheckedResponse):
 
   return;
 }
+
 /** Resets specific documents in the datasource to be selectively re-ingested by the indexer. */
 export async function resetDocuments(
   context: Client,
@@ -934,6 +948,7 @@ export async function _resyncDeserialize(result: PathUncheckedResponse): Promise
 
   return;
 }
+
 /** Resync selective options from the datasource to be re-ingested by the indexer." */
 export async function resync(
   context: Client,
@@ -989,6 +1004,7 @@ export async function _resetIndexerDeserialize(result: PathUncheckedResponse): P
 
   return;
 }
+
 /** Resets the change tracking state associated with an indexer. */
 export async function resetIndexer(
   context: Client,
@@ -1046,6 +1062,7 @@ export async function _createDataSourceConnectionDeserialize(
 
   return searchIndexerDataSourceConnectionDeserializer(result.body);
 }
+
 /** Creates a new datasource. */
 export async function createDataSourceConnection(
   context: Client,
@@ -1104,6 +1121,7 @@ export async function _getDataSourceConnectionsDeserialize(
 
   return _listDataSourcesResultDeserializer(result.body);
 }
+
 /** Lists all datasources available for a search service. */
 export function getDataSourceConnections(
   context: Client,
@@ -1168,6 +1186,7 @@ export async function _getDataSourceConnectionDeserialize(
 
   return searchIndexerDataSourceConnectionDeserializer(result.body);
 }
+
 /** Retrieves a datasource definition. */
 export async function getDataSourceConnection(
   context: Client,
@@ -1226,6 +1245,7 @@ export async function _deleteDataSourceConnectionDeserialize(
 
   return;
 }
+
 /** Deletes a datasource. */
 export async function deleteDataSourceConnection(
   context: Client,
@@ -1289,6 +1309,7 @@ export async function _createOrUpdateDataSourceConnectionDeserialize(
 
   return searchIndexerDataSourceConnectionDeserializer(result.body);
 }
+
 /** Creates a new datasource or updates a datasource if it already exists. */
 export async function createOrUpdateDataSourceConnection(
   context: Client,